fixed my_sort_fn in WellAggFeatureDistanceMetrics
fixed ```my_sort_fn``` not working in the ```WellAggFeatureDistanceMetrics``` class.
Now, the wells are sorted based on "density", "cell-line", "treatment" in the final distancemap excel file output of step5 of cellpaint.
Also, the anchor treatment/DMSO and other controls appear at the beginning rows of excel file. This is very convenient for the biologist.